Public Documentation

Documentation for XRootD.jl public interface.

Index

XRootD.XrdCl.File

XRootD.XrdCl.FileSystem

XRootD.XrdCl.locate

XRootD.XrdCl.ping

XRootD.XrdCl.protocol

XRootD.XrdCl.query

XRootD.XrdCl.rmdir

Base.close

Base.copy

Base.eof

Base.isopen

Base.open

Base.read

Base.readline

Base.readlines

Base.stat

Base.stat

Base.truncate

Base.truncate

Base.write

Types

This is the list of all types and functions defined for XRootD

XRootD.XrdCl.File — Type

File(url::String, flags=0x0000, mode=0x0000)

Create a File object and to open it.

Arguments

url::String: the URL of the file.

flags::Int: the flags to open the file (examples: OpenFlags.Read, OpenFlags.Write, OpenFlags.Delete).

mode::Int: the mode to open the file (examples: Access.UR, Access.UW).

Returns

File: the File object (if the file is opened successfully) or nothing.

XRootD.XrdCl.FileSystem — Type

FileSystem(url::String, isServer::Bool=false)

Creates a FileSystem object. It is a wrapper around the XRootD FileSystem class that is used for sending queries to the XRootD server.

Functions

Base.Filesystem.chmod — Function

Base.chmod(fs::FileSystem, path::String, mode, timeout::UInt16=0x0000)

Change the mode of a file or directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the file or directory

mode: The new mode

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

Base.Filesystem.mkdir — Function

Base.mkdir(fs::FileSystem, path::String, mode::XRootD.XrdCl!Access!Mode=Access.None, timeout::UInt16=0x0000)

Create a directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the directory

mode::XrdCl.Access.Mode: The mode

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

Base.Filesystem.mv — Function

Base.mv(fs::FileSystem, src::String, dest::String, timeout::UInt16=0x0000)

Move or rename a file or directory.

Arguments

fs::FileSystem: The FileSystem object

src::String: The source path

dest::String: The destination path

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

Base.Filesystem.readdir — Function

Base.readdir(fs::FileSystem, path::String, flags::XRootD.XrdCl!DirListFlags!Flags=XRootD.XrdCl!DirListFlags!None; join::Bool = false, sort::Bool = false)

List entries in a directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the directory

flags::XrdCl.DirListFlags: The flags

join::Bool: If true the entries will be joined with the path

sort::Bool: If true the entries will be sorted

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

entries::Vector{String}: The entries in the directory (or nothing if the operation failed)

Base.Filesystem.rm — Function

Base.rm(fs::FileSystem, path::String, timeout::UInt16=0x0000)

Remove a file or directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the file or directory

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

Base.close — Method

Base.close(f::File)

Close the file.

Arguments

f::File: the File object.

Returns

Tuple of:

XRootDStatus: the status of the operation.

nothing

Base.copy — Method

copy(fs::FileSystem, src::String, dest::String; force::Bool=false)

Simple method to copy a file or directory. It uses some default options. For more control use the CopyProcess class.

Arguments

fs::FileSystem: The FileSystem object

src::String: The source path

dest::String: The destination path

force::Bool: If true the destination will be overwritten

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

results::PropertyList: The results of the operation (or nothing if the operation failed)

Base.eof — Method

Base.eof(f::File)

Check if the file is at the end.

Arguments

f::File: the File object.

Returns

Bool: true if the file is at the end, false otherwise.

Base.isopen — Method

Base.isopen(f::File)

Check if the file is open.

Arguments

f::File: the File object.

Returns

Bool: true if the file is open, false otherwise.

Base.open — Function

Base.open(f::File, url::String, flags=0x0000, mode=0x0000)

Open a file.

Arguments

f::File: the File object.

url::String: the URL of the file.

flags::Int: the flags to open the file (examples: OpenFlags.Read, OpenFlags.Write, OpenFlags.Delete).

mode::Int: the mode to open the file (examples: Access.UR, Access.UW).

Returns

Tuple of:

XRootDStatus: the status of the operation.

nothing

Base.read — Function

Base.read(f::File, size, offset=0)

Read data from the file.

Arguments

f::File: the File object.

size::Int: the size of the data to read.

offset::Int: the offset to read the data.

Returns

Tuple of:

XRootDStatus: the status of the operation.

Array{UInt8}: the data read (or nothing if the operation failed).

Base.readline — Function

Base.readline(f::File, size=0, offset=0, chunk=0)

read one line from the file until it finds a linefeed ( ) or reaches the end of the file.

Arguments

f::File: the File object.

size::Int: the maximum size of the line to read.

offset::Int: the offset in the file (0 indicates to continue reading from previous position).

chunk::Int: the size of the chunk to read (0 indicates 2MB).

Returns

Tuple of:

XRootDStatus: the status of the operation.

String: the line read (or nothing if the operation failed).

Base.readlines — Function

Base.readlines(f::File, size=0, offset=0, chunk=0)

readlines reads lines from the file.

Arguments

f::File: the File object.

size::Int: the maximum size of the lines to read.

offset::Int: the offset in the file (0 indicates to continue reading from previous position).

chunk::Int: the size of the chunk to read (0 indicates 2MB).

Returns

Tuple of:

XRootDStatus: the status of the operation.

Array{String}: the lines read (or nothing if the operation failed).

Base.stat — Function

Base.stat(fs::FileSystem, path::String, timeout::UInt16=0x0000)

Gets the StatInfo of a file or directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the file or directory

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

statinfo::StatInfo: The StatInfo object (or nothing if the operation failed)

Base.stat — Function

Base.stat(f::File, force::Bool=true)

Stat the file.

Arguments

f::File: the File object.

force::Bool: force the stat operation.

Returns

Tuple of:

XRootDStatus: the status of the operation.

StatInfo: the stat information (or nothing if the operation failed).

Base.truncate — Function

Base.truncate(fs::FileSystem, path::String, size::Int64, timeout::UInt16=0x0000)

Truncate a file to a specified size.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the file

size::Int64: The new size

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

Base.truncate — Method

Base.truncate(f::File, size::Int64)

Truncate the file.

Arguments

f::File: the File object.

size::Int64: the size to truncate the file.

Returns

Tuple of:

XRootDStatus: the status of the operation.

nothing

Base.write — Function

Base.write(f::File, data::Array{UInt8}, size, offset=0)

Write data to the file.

Arguments

f::File: the File object.

data::Array{UInt8}: the data to write.

size::Int: the size of the data to be writen.

offset::Int: the offset to write the data.

Returns

Tuple of:

XRootDStatus: the status of the operation.

nothing

XRootD.XrdCl.locate — Function

locate(fs::FileSystem, path::String, flags::XRootD.XrdCl!OpenFlags!Flags, timeout::UInt16=0x0000)

Get the locations of a file or directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the file or directory

flags::XrdCl.OpenFlags: The flags

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

locations::Vector{LocationInfo}: The locations of the file or directory (or nothing if the operation failed)

XRootD.XrdCl.ping — Function

ping(fs::FileSystem)

Check if the server is alive - sync

Arguments

fs::FileSystem: The FileSystem object

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always

XRootD.XrdCl.protocol — Function

protocol(fs::FileSystem, timeout::UInt16=0x0000)

Get the protocol information.

Arguments

fs::FileSystem: The FileSystem object

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

protocol::ProtocolInfo: The protocol information (or nothing if the operation failed)

XRootD.XrdCl.query — Function

query(fs::FileSystem, code::XRootD.XrdCl!QueryCode!Code , arg::String, timeout::UInt16=0x0000)

Query the server.

Arguments

fs::FileSystem: The FileSystem object

code::XrdCl.QueryCode: The query code

arg::String: The argument

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

response::String: The response (or nothing if the operation failed)

XRootD.XrdCl.rmdir — Function

rmdir(fs::FileSystem, path::String, timeout::UInt16=0x0000)

Remove a directory.

Arguments

fs::FileSystem: The FileSystem object

path::String: The path of the directory

timeout::UInt16: The timeout in seconds

Returns

Tuple of:

st::XRootD.XrdCl!Status: The status of the operation

nothing: Always